requirements.txt
Django
djangorestframework
requests
djangorestframework-jwt
psycopg2-binary
neomodel
django_neomodel
pillow
cryptography
coreapi
